started TDD for reading xmls into a slideshow format.
in the future we will develop a simple xml document to be read so that
initial tests pass before creating a PWS slideshow to be read and when
that test is passed, create a slideshow specifically tailored for smart
trolley that will be altered every time we change our system.

In this way the pws capability will be tested whenever the system is
changed.