Work Done: The AllShoppingListsScreen is screwed up, should probably revert back to an old version, but will try fixing it tomorrow
Also trying to add an image to the play/pause button, and adding the
favorites button.

Problems Faced:
Problems with all of the above, but no others.

Work to be Done:
Refactor based on Javadoc
Complete the merge with the PWS Parser (XML Parsing branch) that
Alasdair has started on his branch
Look at UML add-ons to Eclipse